Correções de problemas da funcionabilidade ensala
correções no view, model e controle

view: na opções aparecia um espaço vazio
model: corrigido a a função de apagar que causava FATAL ERRO se submetida a uma atualização de página pelo usuário
controle: corrigido o problema de quando filtrado a tabela e o usuário apaga-se um registro, ocorre-se um falha que impedia que essa ação acontecesse.

por fim é isso uma ótima semana 👍

+++ Model/mEnsalamento.php
@@ -385,6 +385,8 @@ function apagar($cod_ensalamento)
     $consulta = $conecxao->query("SELECT codigo_reserva FROM sgrp.reserva_ensalamento
     WHERE codigo_ensalamento = '$cod_ensalamento' ");
 
+    $dados = '';
+
     if ($consulta->num_rows > 0) 
     {
         $dados = [];
@@ -394,30 +396,32 @@ function apagar($cod_ensalamento)
         }
     }
 
-    // var_dump($dados);
+    if ($dados != null){
 
-    $cod_reserva = $dados['0']['codigo_reserva'];
+        $cod_reserva = $dados['0']['codigo_reserva'];
 
-    // var_dump($cod_reserva);
+        // var_dump($cod_reserva);
 
-    $apagar = $conecxao->query("DELETE FROM sgrp.reserva_ensalamento
-    WHERE  codigo_ensalamento = '$cod_ensalamento' and codigo_reserva = '$cod_reserva'");
+        $apagar = $conecxao->query("DELETE FROM sgrp.reserva_ensalamento
+        WHERE  codigo_ensalamento = '$cod_ensalamento' and codigo_reserva = '$cod_reserva'");
 
-    // var_dump($apagar);
-    // echo '<br>';
+        // var_dump($apagar);
+        // echo '<br>';
 
-    $apagar = $conecxao->query("DELETE FROM sgrp.data_reserva WHERE codigo_reserva = '$cod_reserva' ");
+        $apagar = $conecxao->query("DELETE FROM sgrp.data_reserva WHERE codigo_reserva = '$cod_reserva' ");
 
-    // var_dump($apagar);
-    // echo '<br>';
+        // var_dump($apagar);
+        // echo '<br>';
 
-    $apagar = $conecxao->query("DELETE FROM sgrp.reserva WHERE codigo = '$cod_reserva' ");
+        $apagar = $conecxao->query("DELETE FROM sgrp.reserva WHERE codigo = '$cod_reserva' ");
 
-    // var_dump($apagar);
-    // echo '<br>';
-    
-    $apagar = $conecxao->query("DELETE FROM sgrp.ensalamento WHERE codigo = '$cod_ensalamento'");
+        // var_dump($apagar);
+        // echo '<br>';
+        
+        $apagar = $conecxao->query("DELETE FROM sgrp.ensalamento WHERE codigo = '$cod_ensalamento'");
 
-    // var_dump($apagar);
-    // echo '<br>';
+        // var_dump($apagar);
+        // echo '<br>';
+
+    }
 }// fazer
